Handle item model rendering
This adds item models being rendered correctly. For some reason item
transforms aren't handled automatically for ISmartItemModelS so I
had to wrap it inside of an IPerspectiveAwareModel. This might be a
mistake on my part and there might be a cleaner way of doing it, but for
now it works.