Rewrite NBT handling code to use NBTPersist
Complete rewrite of the way NBT handling is done. Most manual NBT
parsing is now handled by Cyclops' NBTPersist methods, although not all
manual parsing could be removed due to the limitations of said system.
A future rewrite in Cyclops will probably fix this is issue, but for now
this is all I could do.