Add support for cyclops' recipe system
This adds support for recipes. If a fluid group has the property
'hasRecipe' set to true, than it can be crafted. Its recipe can then be
specified in an xml file inside the config folder.

Alternatively, if 'hasDefaultRecipe' is also set to true, a default
recipe will be generated for the given fluid converter.