Remove hasRecipe property
Having both 'hasRecipe' and 'hasDefaultRecipe' was confusing. With this
change predefined items will always be registered for all fluid
converters, so custom recipes will only exist if a user has specified a
custom recipe for a fluid converter. 'hasDefaultRecipe' still exists and
will decide if the default fluid converter recipe should be enabled or
not.